requirements.txt
alembic==1.6.5
click<8.0
Flask==1.1.4
Flask-Migrate==2.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
itsdangerous<2.0
Jinja2<3.0
Mako==1.1.5
MarkupSafe==2.0.1
Pillow==8.3.2
PyJWT==2.1.0
python-dateutil==2.8.2
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.23
Werkzeug<2.0
